Added current state of the data puller.
Currently doesn't output to files, Currently doesn't work properly on file input.
Prompts for user input. Accepts file name, accepts user name, asks if there is an API call limitation, requests number of calls per second.
Asks if it's for Channel followers or if it's for a User's followers.
Currently works fine for a single user.